Change tx behavior of StateVec::with_capacity
If my understanding is correct, it would be very rare for a StateVec to have more than a couple of `tx` instances on it. If a user wanted to initialize the vec with 1000 items they would likely not also mean it to be initialized for it to have 1000 attached listeners.